requirements.txt
openai
faiss-cpu
streamlit
tqdm
sentence-transformers
numpy
pymupdf